+++ enterprise/internal/codeintel/codenav/transport/graphql/gitblob_lsif_data_resolver_test.go
@@ -130,8 +130,7 @@ func TestReferences(t *testing.T) {
 			Line:      10,
 			Character: 15,
 		},
-		ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset},
-		After:          &mockCursor,
+		PagedConnectionArgs: resolverstubs.PagedConnectionArgs{ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset}, After: &mockCursor},
 	}
 
 	if _, err := resolver.References(context.Background(), args); err != nil {
@@ -185,7 +184,7 @@ func TestReferencesDefaultLimit(t *testing.T) {
 			Line:      10,
 			Character: 15,
 		},
-		ConnectionArgs: resolverstubs.ConnectionArgs{},
+		PagedConnectionArgs: resolverstubs.PagedConnectionArgs{},
 	}
 
 	if _, err := resolver.References(context.Background(), args); err != nil {
@@ -231,7 +230,7 @@ func TestReferencesDefaultIllegalLimit(t *testing.T) {
 			Line:      10,
 			Character: 15,
 		},
-		ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset},
+		PagedConnectionArgs: resolverstubs.PagedConnectionArgs{ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset}},
 	}
 
 	if _, err := resolver.References(context.Background(), args); err != ErrIllegalLimit {
@@ -308,7 +307,7 @@ func TestDiagnostics(t *testing.T) {
 
 	offset := int32(25)
 	args := &resolverstubs.LSIFDiagnosticsArgs{
-		ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset},
+		First: &offset,
 	}
 
 	if _, err := resolver.Diagnostics(context.Background(), args); err != nil {
@@ -348,9 +347,7 @@ func TestDiagnosticsDefaultLimit(t *testing.T) {
 		mockOperations,
 	)
 
-	args := &resolverstubs.LSIFDiagnosticsArgs{
-		ConnectionArgs: resolverstubs.ConnectionArgs{},
-	}
+	args := &resolverstubs.LSIFDiagnosticsArgs{}
 
 	if _, err := resolver.Diagnostics(context.Background(), args); err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -391,7 +388,7 @@ func TestDiagnosticsDefaultIllegalLimit(t *testing.T) {
 
 	offset := int32(-1)
 	args := &resolverstubs.LSIFDiagnosticsArgs{
-		ConnectionArgs: resolverstubs.ConnectionArgs{First: &offset},
+		First: &offset,
 	}
 
 	if _, err := resolver.Diagnostics(context.Background(), args); err != ErrIllegalLimit {
