+++ enterprise/internal/codeintel/shared/resolvers/audit_log_column_resolver.go
@@ -6,7 +6,7 @@ type auditLogColumnChangeResolver struct {
 	columnTransition map[string]*string
 }
 
-func NewAuditLogColumnChangeResolver(columnTransition map[string]*string) resolverstubs.AuditLogColumnChangeResolver {
+func NewAuditLogColumnChangeResolver(columnTransition map[string]*string) resolverstubs.AuditLogColumnChange {
 	return &auditLogColumnChangeResolver{columnTransition}
 }
