+++ internal/codeintel/resolvers/codenav.go
@@ -1,0 +1,116 @@
+package resolvers
+
+import (
+	"context"
+
+	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/markdown"
+	"github.com/sourcegraph/sourcegraph/internal/types"
+)
+
+type CodeNavServiceResolver interface {
+	GitBlobLSIFData(ctx context.Context, args *GitBlobLSIFDataArgs) (GitBlobLSIFDataResolver, error)
+}
+
+type GitBlobLSIFDataArgs struct {
+	Repo      *types.Repo
+	Commit    api.CommitID
+	Path      string
+	ExactPath bool
+	ToolName  string
+}
+
+type GitBlobLSIFDataResolver interface {
+	GitTreeLSIFDataResolver
+	ToGitTreeLSIFData() (GitTreeLSIFDataResolver, bool)
+	ToGitBlobLSIFData() (GitBlobLSIFDataResolver, bool)
+
+	Stencil(ctx context.Context) ([]RangeResolver, error)
+	Ranges(ctx context.Context, args *LSIFRangesArgs) (CodeIntelligenceRangeConnectionResolver, error)
+	Definitions(ctx context.Context, args *LSIFQueryPositionArgs) (LocationConnectionResolver, error)
+	References(ctx context.Context, args *LSIFPagedQueryPositionArgs) (LocationConnectionResolver, error)
+	Implementations(ctx context.Context, args *LSIFPagedQueryPositionArgs) (LocationConnectionResolver, error)
+	Hover(ctx context.Context, args *LSIFQueryPositionArgs) (HoverResolver, error)
+}
+
+type LSIFRangesArgs struct {
+	StartLine int32
+	EndLine   int32
+}
+
+type LSIFQueryPositionArgs struct {
+	Line      int32
+	Character int32
+	Filter    *string
+}
+
+type LSIFPagedQueryPositionArgs struct {
+	LSIFQueryPositionArgs
+	PagedConnectionArgs
+	Filter *string
+}
+
+type (
+	CodeIntelligenceRangeConnectionResolver = ConnectionResolver[CodeIntelligenceRangeResolver]
+)
+
+type CodeIntelligenceRangeResolver interface {
+	Range(ctx context.Context) (RangeResolver, error)
+	Definitions(ctx context.Context) (LocationConnectionResolver, error)
+	References(ctx context.Context) (LocationConnectionResolver, error)
+	Implementations(ctx context.Context) (LocationConnectionResolver, error)
+	Hover(ctx context.Context) (HoverResolver, error)
+}
+
+type RangeResolver interface {
+	Start() PositionResolver
+	End() PositionResolver
+}
+
+type PositionResolver interface {
+	Line() int32
+	Character() int32
+}
+
+type (
+	LocationConnectionResolver = PagedConnectionResolver[LocationResolver]
+)
+
+type LocationResolver interface {
+	Resource() GitTreeEntryResolver
+	Range() RangeResolver
+	URL(ctx context.Context) (string, error)
+	CanonicalURL() string
+}
+
+type HoverResolver interface {
+	Markdown() Markdown
+	Range() RangeResolver
+}
+
+type Markdown string
+
+func (m Markdown) Text() string {
+	return string(m)
+}
+
+func (m Markdown) HTML() (string, error) {
+	return markdown.Render(string(m))
+}
+
+type GitTreeLSIFDataResolver interface {
+	Diagnostics(ctx context.Context, args *LSIFDiagnosticsArgs) (DiagnosticConnectionResolver, error)
+}
+
+type (
+	LSIFDiagnosticsArgs          = ConnectionArgs
+	DiagnosticConnectionResolver = PagedConnectionResolverWithCount[DiagnosticResolver]
+)
+
+type DiagnosticResolver interface {
+	Severity() (*string, error)
+	Code() (*string, error)
+	Source() (*string, error)
+	Message() (*string, error)
+	Location(ctx context.Context) (LocationResolver, error)
+}
